Variant calling

+++ src/main.cpp
@@ -101,7 +101,7 @@ int main(int argc, const char *argv[]) {
             }
 
             for(int i = 0; i < population_allele_counts.size(); ++i) {
-                double this_allele_freq = (double)x.second[i][j] / sample_depth;
+                double this_allele_freq = (double)x.second[i][j] / (double)sample_depth;
                 if(this_allele_freq >= args.min_minor_freq) {
                     if(this_nucleotides[i] != fasta_parser.seq[j]) {
                         alts_present_at_pos += this_nucleotides[i];
@@ -131,7 +131,7 @@ int main(int argc, const char *argv[]) {
 
             std::priority_queue< std::pair< double, std::string > > q;
             for(int i = 0; i < population_allele_counts.size(); ++i) {
-                double this_allele_freq = (double)x.second[i][j] / sample_depth;
+                double this_allele_freq = (double)x.second[i][j] / (double)sample_depth;
                 if(this_allele_freq >= args.min_minor_freq) {
                     std::string var_info;
                     if(this_nucleotides[i] == fasta_parser.seq[j]) {
